+++ Modules/Intent.Modules.Blazor.HttpClients/release-notes.md
@@ -1,7 +1,7 @@
 ### Version 4.0.2
 
 - Improvement: Improvements in generated http client code: recommended implementation of `IDisposable pattern`, `JsonSerializerOptions` only added if required, as well as small synatax updates
-- Improvement: Non-nullable properties set in all constructors on `HttpClientRequestException`
+- Improvement: Default values set on non-nullable properties on `HttpClientRequestException`
 - Improvement: MediaType `constant` defined and reused in http client, instead of duplicated string literals
 
 ### Version 4.0.1